machines: start splitting VmOverviewCard into smaller sub-components
VmOverviewCard has gotten too big over time.

This commit moves into seperate functional component the relevant bits
for boot order info.
